+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Roo Code Besiege Changelog
 
-## 0.0.1
-
-- Initial release of Roo Code Besiege
+## 0.0.1 2025-02-09
+    - Initial release
+## 0.0.2 2025-02-13
+    - Add provider Aliyun,
+    - Fix volcengine deepseek v3, baidu deepseek v3
+    - Well, update readme and shop url.